+++ lms/djangoapps/ccx/views.py
@@ -49,13 +49,13 @@ from lms.djangoapps.ccx.utils import (
     get_enrollment_action_and_identifiers,
     parse_date
 )
-from lms.djangoapps.courseware.courses import get_course_by_id
 from lms.djangoapps.courseware.field_overrides import disable_overrides
 from lms.djangoapps.grades.api import CourseGradeFactory
 from lms.djangoapps.instructor.enrollment import enroll_email, get_email_params
 from lms.djangoapps.instructor.views.gradebook_api import get_grade_book_page
 from openedx.core.djangoapps.django_comment_common.models import FORUM_ROLE_ADMINISTRATOR, assign_role
 from openedx.core.djangoapps.django_comment_common.utils import seed_permissions_roles
+from openedx.core.lib.courses import get_course_by_id
 from xmodule.modulestore.django import SignalHandler
 
 log = logging.getLogger(__name__)
