+++ lms/djangoapps/certificates/management/commands/fix_ungraded_certs.py
@@ -9,8 +9,8 @@ from django.core.management.base import BaseCommand
 
 from lms.djangoapps.certificates.api import can_generate_certificate_task
 from lms.djangoapps.certificates.models import GeneratedCertificate
-from lms.djangoapps.courseware import courses
 from lms.djangoapps.grades.api import CourseGradeFactory
+from openedx.core.lib.courses import get_course_by_id
 
 log = logging.getLogger(__name__)
 
@@ -50,7 +50,7 @@ class Command(BaseCommand):
         ungraded = GeneratedCertificate.objects.filter(
             course_id__exact=course_id
         ).filter(grade__exact='')
-        course = courses.get_course_by_id(course_id)
+        course = get_course_by_id(course_id)
         for cert in ungraded:
             if can_generate_certificate_task(cert.user, course_id):
                 log.info(f'{course_id} is using V2 certificates. Certificate will not be regraded for user '
