+++ openedx/core/djangoapps/ccxcon/api.py
@@ -13,8 +13,8 @@ from oauthlib.oauth2 import BackendApplicationClient
 from requests_oauthlib import OAuth2Session
 from rest_framework.status import HTTP_200_OK, HTTP_201_CREATED
 
-from lms.djangoapps.courseware.courses import get_course_by_id
 from openedx.core.djangoapps.models.course_details import CourseDetails
+from openedx.core.lib.courses import get_course_by_id
 from common.djangoapps.student.models import anonymous_id_for_user
 from common.djangoapps.student.roles import CourseInstructorRole
