+++ openedx/core/lib/courses.py
@@ -4,6 +4,7 @@ Common utility functions related to courses.
 
 from django import forms
 from django.conf import settings
+from django.http import Http404
 
 from opaque_keys import InvalidKeyError
 from opaque_keys.edx.locator import CourseKey
@@ -82,3 +83,19 @@ def clean_course_id(model_form, is_required=True):
         raise forms.ValidationError(msg)
 
     return course_key
+
+
+def get_course_by_id(course_key, depth=0):
+    """
+    Given a course id, return the corresponding course descriptor.
+
+    If such a course does not exist, raises a 404.
+
+    depth: The number of levels of children for the modulestore to cache. None means infinite depth
+    """
+    with modulestore().bulk_operations(course_key):
+        course = modulestore().get_course(course_key, depth=depth)
+    if course:
+        return course
+    else:
+        raise Http404("Course not found: {}.".format(str(course_key)))
